Erlang hypotenuse calculator
This is a calculator built in erlang, it utilises two user inputs of the
two sides of a right angled triangle and it calculates the hypotenuse of
that right angled triangle. To use this function we should call the
function heipot(X,Y)   the values X and Y are place holders for your
side values.